Add safety against double rollbacks

+++ src/data/rawScripts/setDataState.sh
@@ -1,3 +1,10 @@
+if [ -d "./src/data/snapshots/state-safety" ]; then
+  echo "Already in a rolled back data state."
+  echo "Cannot set a new state without destroying data."
+  echo "run 'yarn revert-data-state' first"
+  exit 1
+fi
+
 rm -rf ./src/data/snapshots/state-safety
 yarn store-snapshot state-safety
